+++ DP-demo/Function.cpp
@@ -11,9 +11,9 @@ Function::~Function()
 {
 }
 
-float Function::func1(float x) {
+double Function::func1(double x) {
 	return x*x - 2 * x + 1;
 }
-float Function::grad1(float x) {
+double Function::grad1(double x) {
 	return 2 * x - 2;
 }
