+++ DP-demo/Operater.cpp
@@ -12,11 +12,11 @@ Operater::~Operater()
 {
 }
 
-float Operater::gradientDescent(float x_start, float step, functype f, functype g, int epoch, float acc, float mom_dis) {
-	float x = x_start;
-	float y = 0;
-	float grad = 0;
-	float pre_grad = 0;
+double Operater::gradientDescent(double x_start, double step, functype f, functype g, int epoch, double acc, double mom_dis) {
+	double x = x_start;
+	double y = 0;
+	double grad = 0;
+	double pre_grad = 0;
 	for (int i = 0; i < epoch; i++) {
 		grad = g(x);
 		pre_grad = pre_grad * mom_dis + grad * step;
